refactor(cli): refactor extensions and toolbox directories to match the new CLI commands structure
