--- packages/cli/src/extensions/web/configure/deployment.ts
@@ -1,13 +1,0 @@
-import { availableAdapters } from "../../../toolbox/web/configure/deployment/adapters";
-import getInstalledAdapter from "../../../toolbox/web/configure/deployment/get-installed-adapter";
-import installAdapter from "../../../toolbox/web/configure/deployment/install-adapter";
-import isAnyAdapterInstalled from "../../../toolbox/web/configure/deployment/is-any-adapter-installed";
-import removeAdapter from "../../../toolbox/web/configure/deployment/remove-adapter";
-
-export default {
-  availableAdapters,
-  getInstalledAdapter,
-  installAdapter,
-  isAnyAdapterInstalled,
-  removeAdapter,
-};
