+++ packages/cli/src/extensions/web/deployment/configure.ts
@@ -1,0 +1,13 @@
+import { availableAdapters } from "../../../toolbox/web/deployment/configure/adapters";
+import getInstalledAdapter from "../../../toolbox/web/deployment/configure/get-installed-adapter";
+import installAdapter from "../../../toolbox/web/deployment/configure/install-adapter";
+import isAnyAdapterInstalled from "../../../toolbox/web/deployment/configure/is-any-adapter-installed";
+import removeAdapter from "../../../toolbox/web/deployment/configure/remove-adapter";
+
+export default {
+  availableAdapters,
+  getInstalledAdapter,
+  installAdapter,
+  isAnyAdapterInstalled,
+  removeAdapter,
+};
