+++ packages/cli/src/toolbox/web/deployment/configure/get-installed-adapter.ts
