+++ packages/cli/src/toolbox/web/deployment/configure/is-any-adapter-installed.ts
