+++ packages/cli/src/toolbox/web/deployment/configure/remove-adapter.ts
